app.py
from flask import Flask
from flask_sqlalchemy import SQLAlchemy
import os
from dotenv import load_dotenv

load_dotenv()


app = Flask(__name__)
app.config['SECRET_KEY'] = os.getenv('SECRET_KEY')
app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///database.db'

db = SQLAlchemy(app)


@app.route('/', methods=['GET'])
def hello_word():
    return '<h1>Hello world</h1>'


if __name__ == '__main__':
    app.run(debug=True)
